% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MASH.R
\name{MASH}
\alias{MASH}
\title{Calculate Mediation effect for Survival Data}
\usage{
MASH(p = 1/2, exp, med, event, date)
}
\arguments{
\item{p}{A numeric value indicating the proportion of the data to be used for training. For example, 0.7 for 70% of the data.}

\item{exp}{A numeric matrix or vector representing the exposure data.}

\item{med}{A numeric matrix representing the mediation data (covariates).}

\item{event}{A binary vector indicating the occurrence of the event of interest.}

\item{date}{A numeric vector indicating the time to event or censoring.}
}
\value{
A list containing the following components:
\itemize{
  \item \code{rho2w_Yx}: rho2w value for the exposure model.
  \item \code{rho2w_Ym}: rho2w value for the mediation model.
  \item \code{rho2w_Ymx}: rho2w value for the combined exposure and mediation model.
  \item \code{r2w}: Overall model fit statistic combining the previous metrics.
  \item \code{SOSw}: A measure of the proportion of the variance explained, specific to the model.
}
}
\description{
This function performs a mediation analysis of survival data based on input data including exposure, mediation, and event data. It uses penalized Cox proportional hazards models and other statistical methods to calculate various metrics, including rho2w and SOSw values.
}
\examples{
# Assuming 'exp', 'med', 'event', and 'date' are already defined:
results <- calculateMetrics(p=0.7, exp=exp, med=med, event=event, date=date)
print(results)
}
