% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData.R
\name{simdata}
\alias{simdata}
\title{Simulate Survival Data}
\usage{
simdata(
  n = n,
  m = m,
  a = a,
  b = bb,
  r = r,
  ppp = ppp,
  c = 1e+08,
  ll = 1e+07,
  seed = 1,
  cov = NULL,
  g = NULL,
  lll = NULL
)
}
\arguments{
\item{n}{Integer, the number of observations (sample size).}

\item{m}{Integer, the number of covariates to generate.}

\item{a}{Numeric, the base value for generating the covariance matrix.}

\item{b}{Numeric vector, coefficients for the covariates.}

\item{r}{Numeric, the coefficient for the primary covariate.}

\item{ppp}{Integer, total number of predictors including covariates.}

\item{c}{Numeric, maximum censoring time.}

\item{ll}{Numeric, maximum follow-up time.}

\item{seed}{Integer, seed for random number generation to ensure reproducibility.}

\item{cov}{Optional matrix, a covariance matrix. If provided, `simdata_cov` is used.}

\item{g}{Optional, additional parameters for `simdata_cov`.}

\item{lll}{Optional, additional parameters for `simdata_cov`.}
}
\value{
A list containing:
  \itemize{
    \item \code{dat}: A matrix with the simulated dataset including survival times, covariates, and event indicators.
    \item \code{MT}: Matrix of covariates.
    \item \code{llt}: Proportion of observations with survival times at maximum follow-up time.
    \item \code{cr}: Censoring rate.
  }
}
\description{
This function simulates survival data based on specified parameters. It can generate data
with or without a covariance matrix. When a covariance matrix is provided, it delegates
the data generation to `simdata_cov`. Otherwise, it creates a covariance matrix among
all covariates and generates survival times, censoring times, and event indicators.
The function also supports generating correlated covariate vectors and adjusting the
scale and shape parameters for survival time calculation.
}
\examples{
simulated_data <- simdata(n=100, m=2, a=0.5, b=c(0.3, 0.5), r=1.2, ppp=5, seed=123)
str(simulated_data)
}
